Wrong language in order status mails
Currently the language ID chain is taken from the sales channel. As there are no subshops like in SW5, the sales channel doesn't have the correct information if an order was places in a different than the default language of this sales channel.
This PR fetches the order context from the current order and passes it instead of the default context.